+++ Lagrange.Core/Internal/Event/Message/GetC2cMessageEvent.cs
@@ -6,13 +6,13 @@ internal class GetC2cMessageEvent : ProtocolEvent
 {
     public string FriendUid { get; } = string.Empty;
 
-    public uint StartSequence { get; }
+    public ulong StartSequence { get; }
 
-    public uint EndSequence { get; }
+    public ulong EndSequence { get; }
 
     public List<MessageChain> Chains { get; } = new();
 
-    private GetC2cMessageEvent(string friendUid, uint startSequence, uint endSequence) : base(true)
+    private GetC2cMessageEvent(string friendUid, ulong startSequence, ulong endSequence) : base(true)
     {
         FriendUid = friendUid;
         StartSequence = startSequence;
@@ -24,7 +24,7 @@ internal class GetC2cMessageEvent : ProtocolEvent
         Chains = chains;
     }
 
-    public static GetC2cMessageEvent Create(string friendUid, uint startSequence, uint endSequence)
+    public static GetC2cMessageEvent Create(string friendUid, ulong startSequence, ulong endSequence)
         => new(friendUid, startSequence, endSequence);
 
     public static GetC2cMessageEvent Result(int retcode, List<MessageChain> chains)
